tests: Reduce duplication of master setup code
